Add parameters to use different keys and repos.
Now, from the command line you can set the key to use as well as the
repository.
Show at start the value of the parameters to be sure that we set the
options as we needed.
Add the option to use a repo file instead of downloading the latest one,
this is useful for the first time release.