Add fabric file to update the TUF repo.
Add support for json config file.
This feature looks like it will be included soon in newer versions of
fabric, but right now we add it manually.

Add the configuration into a json file to avoid data on the python file.

Ignore the file fabfile.json since it contains server information that
may be sensitive.