[feat] Detect sumo top folder with tar
The script was assuming that the sumo tarball was named after the
version of bitmask it contains, but "latest" symlink generated by
buildbot doesn't follow that rule.

Let tar list the folder the tarball generates (assuming there is only
one folder, and that it's listed in first place), and follow accordingly.

- Related: #7345